Implement recursion
Implement recursion by adding an implicit self-reference inside `func`
forms bound to the symbol `recur`.

It requires using a local, mutable atom as the environment that the
function closes over, but this environment itself is lexically scoped
to the `make-fn` function, so we're unlikely to run into any
issues. :fingers-crossed: